added changes to RemindController, GET & DELETE requests are working …

+++ src/main/java/com/muzkat/reminder/service/RemindService.java
@@ -63,6 +63,9 @@ public class RemindService {
      * @param id - идентификатор напоминания
      */
     public void deleteRemind(Long id) {
+        if (!remindRepository.existsById(id)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Напоминание с id " + id + " не найдено");
+        }
         remindRepository.deleteById(id);
     }
